Flytt felt util ut i felles, sammen med test
