+++ src/frontend/components/SøknadsSteg/OmBarnaDine/utils.ts
@@ -3,32 +3,7 @@ import { Felt, ISkjema } from '@navikt/familie-skjema';
 
 import { IBarn } from '../../../typer/person';
 import { ISøknad } from '../../../typer/søknad';
-import { BarnetsIdent } from './HvilkeBarnCheckboxGruppe';
 import { IOmBarnaDineFeltTyper } from './useOmBarnaDine';
-
-export type JaNeiSpmOgCheckboxPar = {
-    jaNeiSpm: Felt<ESvar | undefined>;
-    checkbox: Felt<BarnetsIdent[]>;
-}[];
-
-export const hentFiltrerteAvhengigheter = (jaNeiSpmOgCheckboxPar: JaNeiSpmOgCheckboxPar) => {
-    let avhengigheter = {};
-
-    jaNeiSpmOgCheckboxPar.forEach(par => {
-        if (par.jaNeiSpm.verdi === ESvar.JA) {
-            avhengigheter = {
-                ...avhengigheter,
-                [par.checkbox.id]: par.checkbox,
-            };
-        } else {
-            avhengigheter = {
-                ...avhengigheter,
-                [par.jaNeiSpm.id]: par.jaNeiSpm,
-            };
-        }
-    });
-    return avhengigheter;
-};
 
 export const genererSvarForSpørsmålBarn = (barn: IBarn, felt: Felt<string[]>): ESvar =>
     felt.verdi.includes(barn.ident) ? ESvar.JA : ESvar.NEI;
